feat(runtime, react): DRY test helper and move helper function to utils
