+++ packages/v2/shared/src/index.ts
@@ -8,11 +8,7 @@ export {
 export * from "./utils";
 
 export { logger } from "./logger";
-export {
-  DEFAULT_AGENT_ID,
-  AG_UI_CHANNEL_EVENT,
-  phoenixExponentialBackoff,
-} from "./constants";
+export { DEFAULT_AGENT_ID, AG_UI_CHANNEL_EVENT } from "./constants";
 export { finalizeRunEvents } from "./finalize-events";
 
 export {
